+++ dashboard/models/__init__.py
@@ -1,0 +1,15 @@
+from models.data_models.award_model import Award
+from models.data_models.book_chapter_model import BookChapter
+from models.data_models.book_model import Book
+from models.data_models.classes_model import Classes
+from models.data_models.conference_presentation_model import (
+    ConferencePresentation,
+)
+from models.data_models.journal_model import Journal
+from models.data_models.journal_paper_model import JournalPaper
+from models.data_models.person_model import Person
+from models.data_models.project_model import Project
+from models.data_models.publication_model import Publication
+from models.data_models.research_group_model import ResearchGroup
+from models.data_models.subject_model import Subject
+from models.data_models.thesis_model import Thesis
