+++ dashboard/models/data_models/subject_model.py
@@ -1,0 +1,11 @@
+from models.custom_model import CustomModel
+
+
+class Subject(CustomModel):
+    subject: str
+    career: str
+    semester: int
+    year: int
+
+    def __str__(self) -> str:
+        return self.subject
